backend/requirements.txt
fastapi
uvicorn
python-dotenv
alpaca-py
spotipy
discord.py
openai
pyttsx3
whisper
ffmpeg-python

